+++ test/setup.js
@@ -8,6 +8,9 @@ const chai = require('chai')
 const path = require('path')
 const mod = require('module')
 const sinon = require('sinon')
+const sinonChai = require('sinon-chai')
+
+chai.use(sinonChai)
 
 global.should = chai.should()
 global.sinon = sinon